--- resources/js/components/AppSidebar.vue
@@ -1,38 +1,0 @@
-<script setup lang="ts">
-import { mainNavLinks, rightNavLinks } from '@/layouts/app-layout/nav-links';
-import { dashboard } from '@/routes';
-
-import { Sidebar, SidebarContent, SidebarFooter, SidebarHeader, SidebarMenu, SidebarMenuButton, SidebarMenuItem } from '@/components/ui/sidebar';
-import { Link } from '@inertiajs/vue3';
-
-import NavFooter from '@/components/NavFooter.vue';
-import NavMain from '@/components/NavMain.vue';
-import NavUser from '@/components/NavUser.vue';
-import AppLogo from './AppLogo.vue';
-</script>
-
-<template>
-    <Sidebar collapsible="icon" variant="sidebar">
-        <SidebarHeader>
-            <SidebarMenu>
-                <SidebarMenuItem>
-                    <SidebarMenuButton size="lg" as-child>
-                        <Link :href="dashboard()">
-                            <AppLogo />
-                        </Link>
-                    </SidebarMenuButton>
-                </SidebarMenuItem>
-            </SidebarMenu>
-        </SidebarHeader>
-
-        <SidebarContent>
-            <NavMain :items="mainNavLinks" />
-        </SidebarContent>
-
-        <SidebarFooter>
-            <NavFooter :items="rightNavLinks" />
-            <NavUser />
-        </SidebarFooter>
-    </Sidebar>
-    <slot />
-</template>
