--- resources/js/components/AppSidebarHeader.vue
@@ -1,27 +1,0 @@
-<script setup lang="ts">
-import Breadcrumbs from '@/components/Breadcrumbs.vue';
-import { SidebarTrigger } from '@/components/ui/sidebar';
-import { TBreadcrumb } from '@/types';
-
-withDefaults(
-    defineProps<{
-        breadcrumbs?: TBreadcrumb[];
-    }>(),
-    {
-        breadcrumbs: () => [],
-    },
-);
-</script>
-
-<template>
-    <header
-        class="flex h-16 shrink-0 items-center gap-2 border-b border-sidebar-border/70 px-6 transition-[width,height] ease-linear group-has-data-[collapsible=icon]/sidebar-wrapper:h-12 md:px-4"
-    >
-        <div class="flex items-center gap-2">
-            <SidebarTrigger class="-ml-1" />
-            <template v-if="breadcrumbs && breadcrumbs.length > 0">
-                <Breadcrumbs :breadcrumbs="breadcrumbs" />
-            </template>
-        </div>
-    </header>
-</template>
