+++ resources/js/components/elements/AppLogoIcon.vue
