+++ resources/js/components/elements/DeleteUser.vue
