+++ resources/js/components/elements/PlaceholderPattern.vue
