+++ resources/js/components/elements/TwoFactorSetupModal.vue
