+++ resources/js/layouts/app-layout/AppContent.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
-import { SidebarInset } from '@/components/ui/sidebar';
 import { computed } from 'vue';
+
+import { SidebarInset } from '@/components/ui/sidebar';
 
 interface Props {
     variant?: 'header' | 'sidebar';
