+++ resources/js/layouts/app-layout/AppLogo.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import AppLogoIcon from '@/components/AppLogoIcon.vue';
+import AppLogoIcon from '@/components/elements/AppLogoIcon.vue';
 </script>
 
 <template>
