+++ resources/js/layouts/app-layout/AppShell.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
-import { SidebarProvider } from '@/components/ui/sidebar';
 import { usePage } from '@inertiajs/vue3';
+
+import { SidebarProvider } from '@/components/ui/sidebar';
 
 interface Props {
     variant?: 'header' | 'sidebar';
