+++ resources/js/layouts/app-layout/UserMenuContent.vue
@@ -1,11 +1,12 @@
 <script setup lang="ts">
-import UserInfo from '@/components/UserInfo.vue';
 import { DropdownMenuGroup, DropdownMenuItem, DropdownMenuLabel, DropdownMenuSeparator } from '@/components/ui/dropdown-menu';
 import { logout } from '@/routes';
 import { edit } from '@/routes/profile';
 import type { TUser } from '@/types/models';
 import { Link, router } from '@inertiajs/vue3';
 import { LogOut, Settings } from 'lucide-vue-next';
+
+import UserInfo from './UserInfo.vue';
 
 interface Props {
     user: TUser;
