+++ resources/js/layouts/auth-layout/AuthCardLayout.vue
@@ -1,8 +1,10 @@
 <script setup lang="ts">
-import AppLogoIcon from '@/components/AppLogoIcon.vue';
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
 import { home } from '@/routes';
+
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
 import { Link } from '@inertiajs/vue3';
+
+import AppLogoIcon from '@/components/elements/AppLogoIcon.vue';
 
 defineProps<{
     title?: string;
