+++ resources/js/layouts/auth-layout/AuthSimpleLayout.vue
@@ -1,7 +1,9 @@
 <script setup lang="ts">
-import AppLogoIcon from '@/components/AppLogoIcon.vue';
 import { home } from '@/routes';
+
 import { Link } from '@inertiajs/vue3';
+
+import AppLogoIcon from '@/components/elements/AppLogoIcon.vue';
 
 defineProps<{
     title?: string;
