+++ resources/js/layouts/auth-layout/AuthSplitLayout.vue
@@ -1,7 +1,10 @@
 <script setup lang="ts">
-import AppLogoIcon from '@/components/AppLogoIcon.vue';
 import { home } from '@/routes';
-import { Link, usePage } from '@inertiajs/vue3';
+import { usePage } from '@inertiajs/vue3';
+
+import { Link } from '@inertiajs/vue3';
+
+import AppLogoIcon from '@/components/elements/AppLogoIcon.vue';
 
 const page = usePage();
 const name = page.props.name;
