+++ resources/js/layouts/auth-layout/index.ts
@@ -1,0 +1,1 @@
+export { default as AuthLayout } from './Layout.vue';
