+++ resources/js/lib/utils.ts
@@ -1,5 +1,7 @@
-import { InertiaLinkProps } from '@inertiajs/vue3';
-import { clsx, type ClassValue } from 'clsx';
+import type { InertiaLinkProps } from '@inertiajs/vue3';
+import type { ClassValue } from 'clsx';
+
+import { clsx } from 'clsx';
 import { twMerge } from 'tailwind-merge';
 
 export function cn(...inputs: ClassValue[]) {
